Add testing and refactoring

+++ terraform/scripts/setup_lambda_package.sh
@@ -1,18 +1,9 @@
-#!/usr/bin/env bash
-#
-#  Prepare the Lambda function deployment package by installing the dependencies
-# relative to the function. Terraform will build the package zip and will deploy
-# it to AWS.
-##
+#!/usr/bin/env bash -x
 
 TMP=/tmp/dockerfilegenerator
 
-if [ -d $TMP ]; then
-  rm -rf $TMP;
-fi
-
+if [ -d $TMP ]; then rm -rf $TMP; fi
 mkdir $TMP
 cp ../lambda_function.py $TMP
-for mod in $(cat ../requirements.txt); do
-  pip install $mod -t $TMP;
-done;
+cp -r ../dockerfilegenerator $TMP
+pip install -r ../requirements.txt --system -t $TMP
